Fixed issue with some cards staying flipped
It was a problem with how you are unsetting them. After a little research, it seems that the problem was related with setTimeout and setInterval [https://javascript.info/settimeout-setinterval]. Your code allowed to schedule additional callbacks (functions which are called after the delay) even though ones were already scheduled. This was corrupting the internal state of the problem because your toggledCards list was not longer correct, and some cards were remaining toggled.

I added a boolean var checkingForMatch, which blocks card toggles when set (and also blocks additional checkForMatch calls). It was important to unset checkingForMatch after your 1 second delay in the case where there is no match, because the function really isn't done until the cards have been unset, which is technically happening asynchronously.

In other words setTimeout/setInterval are asynchronous functions that will schedule work, return immediately, and do that work later on after some delay. Without these kinds of functions, your program would freeze for 1 second while it waits to continue.